大幅度优化系统用户列表

+++ matrix-manager/src/main/java/com/matrix/service/impl/McUserInfoServiceImpl.java
@@ -236,7 +236,10 @@ public class McUserInfoServiceImpl extends BaseServiceImpl<McUserInfo, Integer> 
 		try {
 			McUserInfoView view = mcUserInfoDao.loadUserInfo(id);
 			int count = mcUserInfoDao.deleteById(id);   
-			int count_ = mcUserRoleDao.deleteByUserId(id);
+			int count_ = 1;
+			if(StringUtils.isNotBlank(launch.loadDictCache(DCacheEnum.McUserRole , "InitMcUserRole").get(id.toString()))) {
+				count_ = mcUserRoleDao.deleteByUserId(id); // 确定该用户有角色被分配才去删除
+			}
 			int cout__ = mcUserInfoExtDao.deleteByUserId(id);  // 删除mc_user_info_ext表的用户扩展信息 
 			if(count == 1 && count_ == 1 && cout__ == 1){
 				launch.loadDictCache(DCacheEnum.McUserRole , null).del(id.toString());
